// web-frontend/src/components/Canvas/types.ts
export interface Grip {
  x: number;
  y: number;
  side: "top" | "bottom" | "left" | "right";
}
import { KonvaEventObject } from "konva/lib/Node";

export interface ComponentItem {
  id: number; // Optional because library items don't have ids
  name: string;
  icon: string; // Required, not optional
  svg: string; // Required, not optional
  class: string;
  object: string;
  args: any[];
  grips?: Grip[];
  isCustom?: boolean;
  // Additional properties from first block
  legend?: string;
  suffix?: string;
  description?: string;
  png?: string;
}
export interface CanvasItem extends ComponentItem {
  id: number;
  x: number;
  component_id: number; // This is a foreing and primary key
  y: number;
  width: number;
  height: number;
  rotation: number;
  sequence: number; // increasing counter for insertion order
  addedAt: number; // timestamp
  label?: string; // e.g. PRV01A/B or Insulation01
  objectKey?: string; // for counting
  naturalWidth?: number;
  naturalHeight?: number;
}